Have main.dart take a SendPort as a parameter
This allows spawnuri-client to be run from the command line or reply
to the request via a SendPort.

Enhanced the test so it tests spawnUri. I was able to valiate that it
works but I couldn't make it work in a test. The test would just exit
without running any tests.